DeepLearning(Sem4)/requirements.txt
matplotlib==3.6.3
numpy==1.23.5
pandas==1.5.3
scikit_learn==1.2.0
tensorflow_datasets==4.9.2
tensorflow_intel==2.13.0
yfinance==0.2.53
